requirements.txt

psutil==6.0.0
tabulate==0.9.0
termcolor
requests
PyYAML
